+++ src/idea-details/component/idea-details.jsx
@@ -5,7 +5,7 @@ import OpenStadComponentLibs from '../../libs/index.jsx';
 import OpenStadComponentPoll from '../../poll/index.jsx';
 import OpenStadComponentReactions from '../../reactions/index.jsx';
 
-import VoteButton from './vote-button.jsx';
+import VoteButtons from './vote-buttons.jsx';
 import { IdeaImage as OpenStadComponentIdeaImage } from '../../image/index.jsx';
 
 'use strict';
@@ -37,6 +37,7 @@ export default class IdeasDetails extends OpenStadComponent {
       poll: {
         canAddPolls: false,
       },
+      vote: {},
       labels: {},
       types: null,
       typeField: 'typeId',
@@ -253,11 +254,11 @@ export default class IdeasDetails extends OpenStadComponent {
     }
 
     let voteButtonsHTML = null;
-    if (self.config.idea.showVoteButtons) {
+    if (self.config.idea.showVoteButtons && ( typeof self.config.vote.voteValues == 'undefined' || self.config.vote.voteValues.length > 0 )) {
       voteButtonsHTML = (
         <div className="osc-vote-buttons-container">
           <h3>Likes</h3>
-          <VoteButton config={{ caption: 'eens', opinion: 'yes', api: this.config.api, user: this.config.user, siteId: this.config.siteId }} idea={this.state.idea} name="likebutton" value={idea.yes}/>
+          <VoteButtons config={this.config} idea={this.state.idea} name="likebutton"/>
         </div>
       );
     }
@@ -300,10 +301,11 @@ export default class IdeasDetails extends OpenStadComponent {
       let config = {...self.config}
       config.argument.isActive = this.config.argument.isActive && !this.config.argument.ignoreReactionsForIdeaIds.match(new RegExp(`(?:^|\\D)${idea.id}(?:\\D|$)`));
       config.argument.isClosed = this.config.argument.isClosed || this.config.argument.closeReactionsForIdeaIds.match(new RegExp(`(?:^|\\D)${idea.id}(?:\\D|$)`));
+      let loginUrl = OpenStadComponentLibs.auth.getLoginUrl(self.config);
       reactionsHTML = (
         <div>
 			    <div id="reactions" className="osc-reactions-header"><h3>{self.config.argument.title || 'Reacties'}</h3></div>
-          <OpenStadComponentReactions config={{ ...self.config.argument, title: undefined, api: self.config.api, user: self.config.user, siteId: self.config.siteId, ideaId: idea.id, loginUrl: self.config.loginUrl, }}/>
+          <OpenStadComponentReactions config={{ ...self.config.argument, title: undefined, api: self.config.api, user: self.config.user, siteId: self.config.siteId, ideaId: idea.id, loginUrl: loginUrl, }}/>
         </div>
       );
     }
@@ -328,7 +330,7 @@ export default class IdeasDetails extends OpenStadComponent {
 
     }
 
-    let authorHTML = idea.user.nickName || idea.user.fullName || idea.user.firstName +' ' + idea.user.lastName;
+    let authorHTML = idea.user.displayName;
     if (this.config.linkToUserPageUrl) {
       authorHTML = <a href={this.config.linkToUserPageUrl + '/' + idea.user.id} className="osc-author-link">{authorHTML}</a>
     }