+++ src/ideas-on-map/component/idea-form.jsx
@@ -31,6 +31,13 @@ export default class IdeasForm extends OpenStadComponent {
 
     let fields = self.config.idea.fields = [ ...self.config.idea.fields ];
 
+    fields.forEach((field) => {
+		  if ( field.inputType ==	 "image-upload" ) {
+			  field.allowMultiple = self.config.image && self.config.image.allowMultipleImages;
+		  }
+	  });
+    
+
     if (!self.props.idea.extraData) self.props.idea.extraData = {};
 
     let state = {
@@ -288,7 +295,7 @@ export default class IdeasForm extends OpenStadComponent {
 					  <h2>
               Naam
 					  </h2>
-            {self.state.formfields.user && self.state.formfields.user.fullName}
+            {self.state.formfields.user && self.state.formfields.user.displayName}
           </div>
 
           <div className="osc-form-group">